feat: remove shitty map navigation feature
